backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
requests
clerk-sdk-python